requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
asttokens==2.4.1
attrs==24.2.0
bitsets==0.8.4
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
concepts==0.9.2
coverage==7.6.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.0.1
Faker==27.0.0
features==0.5.12
fileconfig==0.6.1
graphviz==0.20.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.1
matplotlib-inline     0.1.7
nose                  1.3.7
openpyxl              3.1.5
outcome               1.3.0.post0
packaging             24.1
parso                 0.8.4
pip                   21.2.4
pluggy                1.5.0
prompt_toolkit        3.0.47
pure_eval             0.2.3
py                    1.11.0
pycparser             2.22
Pygments              2.18.0
PySocks               1.7.1
pytest                7.4.4
pytest-cov            5.0.0
pytest-forked         1.6.0
pytest-html           3.2.0
pytest-metadata       3.1.1
pytest-sugar          1.0.0
pytest-testrail       2.9.0
pytest-xdist          2.5.0
python-dateutil       2.9.0.post0
python-dotenv         1.0.1
pytz                  2024.1
requests              2.32.3
selenium              4.10.0
setuptools            58.1.0
six                   1.16.0
sniffio               1.3.1
sortedcontainers      2.4.0
stack-data            0.6.3
termcolor             2.4.0
tomli                 2.0.1
tqdm                  4.66.5
traitlets             5.14.3
trio                  0.26.2
trio-websocket        0.11.1
typing_extensions     4.12.2
urllib3               2.2.2
wcwidth               0.2.13
webdriver-manager     4.0.2
websocket-client      1.8.0
wsproto               1.2.0
